Update workflow actions and dependency tooling
- bump GitHub Actions versions in CI workflow
- refresh CMake presets and dependency CMake modules
- rename dep-toolchain-clang.cmake to dep-tools-clang.cmake
- apply related compatibility fixes in lwIP sys_arch and ds18b20

+++ Core/App/Sensors/ds18b20/ds18b20.cpp
@@ -474,8 +474,8 @@ void DS18B20Sensor::handleSearchTransaction(void)
 				}
 			}
 
-			uint8_t ByteIdx = (_SearchBitNumber - 1U) / BITS_IN_BYTE;
-			uint8_t BitIdx = (_SearchBitNumber - 1U) % BITS_IN_BYTE;
+			uint8_t ByteIdx = static_cast<uint8_t>((_SearchBitNumber - 1U) / BITS_IN_BYTE);
+			uint8_t BitIdx = static_cast<uint8_t>((_SearchBitNumber - 1U) % BITS_IN_BYTE);
 			if(SearchDirection != 0U)
 			{
 				_SearchRomCode[ByteIdx] |= static_cast<uint8_t>(1U << BitIdx);
